Improve connection an account manager component

+++ projects/accounts-manager/src/accounts-manager.ts
@@ -1,43 +1,27 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { Injectable, Injector } from '@angular/core';
-import { AlertType, IBaseItem, WebSocketService, WSEventType } from 'communication';
+import { AlertType, Id, WebSocketService, WSEventType } from 'communication';
 import { NotificationService } from 'notification';
 // Todo: Make normal import
 // The problem now - circular dependency 
 import { accountsListeners } from '../../real-trading/src/connection/accounts-listener';
-import { Observable, of, throwError } from 'rxjs';
-import { catchError, concatMap, map, tap } from 'rxjs/operators';
+import { BehaviorSubject, Observable, of, throwError } from 'rxjs';
+import { catchError, concatMap, map, mergeMap, tap } from 'rxjs/operators';
 import { AccountRepository, ConnectionsRepository, IAccount, IConnection } from 'trading';
-import { AccountNode, AccountNodeSubscriber, IAccountNodeData } from './account.node';
 
 @Injectable()
 export class AccountsManager {
-  private _connectionsData = this._getAccountNodeData();
-  private __connections: IConnection[] = [];
+
+  // private __connections: IConnection[] = [];
 
   private get _connections(): IConnection[] {
-    return this.__connections;
+    return this.connectionsChange.value;
   }
 
   private set _connections(value: IConnection[]) {
-    this._connectionsData = this._getAccountNodeData(this._connections, value);
-    this.__connections = value;
-    this._connectedConnections = value.filter(i => i.connected);
+    this.connectionsChange.next(value);
   }
 
-  private _connectedConnectionsData = this._getAccountNodeData();
-  private __connectedConnections: IConnection[] = [];
-
-  private get _connectedConnections(): IConnection[] {
-    return this.__connectedConnections;
-  }
-
-  private set _connectedConnections(value: IConnection[]) {
-    this._connectedConnectionsData = this._getAccountNodeData(this._connectedConnections, value);
-    this.__connectedConnections = value;
-  }
-
-  private _accountsData = this._getAccountNodeData<IAccount>();
   private __accounts: IAccount[] = [];
 
   private get _accounts(): IAccount[] {
@@ -45,14 +29,15 @@ export class AccountsManager {
   }
 
   private set _accounts(value: IAccount[]) {
-    this._accountsData = this._getAccountNodeData<IAccount>(this._accounts, value);
     this.__accounts = value;
   }
 
-  private _subscribers = new Map<AccountNode, Set<AccountNodeSubscriber>>();
+
 
   private _wsIsOpened = false;
   private _wsHasError = false;
+
+  connectionsChange = new BehaviorSubject<IConnection[]>([]);
 
   constructor(
     protected _injector: Injector,
@@ -68,22 +53,25 @@ export class AccountsManager {
     // this._interceptor.disconnectError.subscribe(() => this._deactivateConnection());
 
     await this._fetchConnections();
-    await this._emitData();
 
     for (const connection of this._connections) {
       if (!connection.connected)
         continue;
 
-      this._wsInit(connection);
+      this._initWS(connection);
       this._fetchAccounts(connection);
     }
 
     return this._connections;
   }
 
-  _fetchAccounts(connection) {
+  private _fetchAccounts(connection: IConnection) {
     const repo = this._accountRepository.get(connection);
-    repo.getItems().subscribe(console.log);
+
+    this._getAccountsByConnections(connection).then(accounts => {
+      this._accounts = this._accounts.concat(accounts);
+      accountsListeners.notifyAccountsConnected(accounts, this._accounts);
+    })
   }
 
   // subscribe(node: AccountNode, ...subscribers: AccountNodeSubscriber[]) {
@@ -125,6 +113,7 @@ export class AccountsManager {
           item.connected = false;
           delete item.connectionData;
         }
+
         return item;
       });
     });
@@ -162,7 +151,7 @@ export class AccountsManager {
       .toPromise().then((i) => i.data);
   }
 
-  private _wsInit(connection: IConnection) {
+  private _initWS(connection: IConnection) {
     const webSocketService = this._webSocketService.get(connection);
 
     webSocketService.on(WSEventType.Message, this._wsHandleMessage.bind(this));
@@ -174,7 +163,7 @@ export class AccountsManager {
     webSocketService.send({ type: 'Id', value: connection.connectionData.apiKey });
   }
 
-  private _wsClose(connection: IConnection) {
+  private _closeWS(connection: IConnection) {
     this._webSocketService.destroy(connection);
   }
 
@@ -237,7 +226,7 @@ export class AccountsManager {
     const _connection = { ...connection, name };
 
     return this._connectionsRepository.updateItem(_connection)
-      .pipe(tap(() => this.onUpdated(_connection, false)));
+      .pipe(tap(() => this.onUpdated(_connection)));
   }
 
   connect(connection: IConnection): Observable<IConnection> {
@@ -249,10 +238,11 @@ export class AccountsManager {
         }),
         tap((conn) => {
           if (conn.connected) {
-            this._wsInit(conn);
+            this._initWS(conn);
+            this._fetchAccounts(connection);
           }
 
-          this.onUpdated(conn, !conn.error);
+          this.onUpdated(conn);
         }),
       );
   }
@@ -261,6 +251,7 @@ export class AccountsManager {
     const disconectedAccounts = this._accounts.filter(account => account.connectionId === connection.id);
     this._accounts = this._accounts.filter(account => account.connectionId === connection.id);
     accountsListeners.notifyAccountsConnected(disconectedAccounts, this._accounts);
+    this._closeWS(connection);
   }
 
   disconnect(connection: IConnection): Observable<void> {
@@ -283,9 +274,16 @@ export class AccountsManager {
   deleteConnection(connection: IConnection): Observable<any> {
     const { id } = connection;
 
-    return this._connectionsRepository.deleteItem(id)
+    return (connection.connected ? this.disconnect(connection) : of(null))
       .pipe(
-        tap(() => this.onDeleted({ id } as IConnection)),
+        mergeMap(() => this._connectionsRepository.deleteItem(id)),
+        catchError((error) => {
+          if (error.status === 401)
+            return this._connectionsRepository.deleteItem(id);
+
+          return throwError(error);
+        }),
+        tap(() => this._connections = this._connections.filter(i => i.id !== id)),
       );
   }
 
@@ -294,51 +292,20 @@ export class AccountsManager {
 
     return this._connectionsRepository.updateItem(_connection)
       .pipe(
-        tap(() => this.onUpdated(_connection, false)),
+        tap(() => this.onUpdated(_connection)),
       );
   }
 
   protected onCreated(connection: IConnection): void {
-  //   if (!connection.name) {
-  //     connection.name = `${connection.server}(${connection.gateway})`;
-  //   }
+    if (!connection.name) {
+      connection.name = `${connection.server}(${connection.gateway})`;
+    }
 
     this._connections = this._connections.concat(connection);
-
-    this._emitData();
-  //   this._emitConnection(connection);
   }
 
-  protected onDeleted(connection: IConnection): void {
-    this._connections = this._connections.filter(i => i.id !== connection.id);
-
-    this._emitData();
-  //   this._emitConnection(connection);
-  }
-
-  protected onUpdated(connection: IConnection, emitConnection = true): void {
+  protected onUpdated(connection: IConnection): void {
     this._connections = this._connections.map(i => i.id === connection.id ? connection : i);
-
-    this._emitData();
-
-  //   if (emitConnection) {
-  //     this._emitConnection(connection);
-  //   }
-  }
-
-  private async _emitData(): Promise<void> {
-    const accounts = this._accounts.filter(account => {
-      return this._connectedConnections.some(i => i.id === account.connectionId);
-    });
-
-    const _accounts = await this._getAccountsByConnections(this._connections.filter(i => i.connected)[0]);
-
-    this._accounts = accounts.concat(_accounts);
-    accountsListeners.notifyAccountsConnected(_accounts, this._accounts);
-
-    // this._forEachSubscriber(subscriber => {
-    //   this._emitDataToSubscriber(subscriber);
-    // });
   }
 
   // private _emitDataToSubscriber(subscriber: AccountNodeSubscriber) {
@@ -346,17 +313,6 @@ export class AccountsManager {
   //   subscriber.handleConnectedConnectionsChange(this._connectedConnectionsData);
   //   subscriber.handleAccountsChange(this._accountsData);
   // }
-
-  // private _emitConnection(connection: IConnection, account?: IAccount) {
-  //   const nodes = this._subscribers.keys();
-
-  //   for (let node of nodes) {
-  //     if (connection.id === node.connection?.id && connection.connected !== node.connection?.connected) {
-  //       const _connection = this._connections.find(i => i.id === connection.id);
-
-  //       this._emitConnectionToNode(node, _connection, account);
-  //     }
-  //   }
 
   //   if (!connection.connected) {
   //     this._wsClose(connection);
@@ -398,21 +354,4 @@ export class AccountsManager {
   //   }
   // }
 
-  private _getAccountNodeData<T extends IBaseItem = IConnection>(prev: T[] = [], current: T[] = []): IAccountNodeData<T> {
-    const deleted = prev.filter(i => !current.some(_i => _i.id === i.id));
-    const created = current.filter(i => !prev.some(_i => _i.id === i.id));
-
-    return {
-      current,
-      prev,
-      created,
-      deleted,
-    };
-  }
-
-  private _forEachSubscriber(callback: (subscriber: AccountNodeSubscriber, node: AccountNode) => void) {
-    this._subscribers.forEach((subscribers, node) => {
-      subscribers.forEach(subscriber => callback(subscriber, node));
-    });
-  }
 }
